[#187] modify: 각 필드 데이터 올바른 타입으로 수정 (#189)
* [#187] fix: 각 필드 데이터 올바른 타입으로 수정

* [#187] modify: Mileage 타입 Integer -> Double

+++ rest/src/main/java/com/wherecar/rest/car/domain/CarStatus.java
@@ -18,14 +18,14 @@ public class CarStatus  extends BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name="car_status_id")
-    private long id;
+    private Long id;
 
     @OneToOne(fetch = FetchType.LAZY)
     @JoinColumn(name="car_id")
     private Car car;
 
     @Column(name="mileage")
-    private Integer mileage;
+    private Double mileage;
 
     @Column(name="battery_voltage")
     private Integer batteryVoltage;
