+++ rest/src/main/java/com/wherecar/rest/carlog/application/dto/CarLogResponse.java
@@ -26,9 +26,9 @@ public class CarLogResponse {
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd HH:mm:ss")
     private LocalDateTime offTime;
 
-    private Integer onMileage;
-    private Integer offMileage;
-    private Integer totalMileage;
+    private Double onMileage;
+    private Double offMileage;
+    private Double totalMileage;
     private DriveType driveType;
     private String driver;
     private String description;
